Move failed test cases painting code to the CloudLayouterPainter class
